Build 20072901
- Changed format version varables
- Added basic serial API functions
- Dropped ilib280 support (depreciated)
- Added Adafruit BME280 support
- Corrected datatypes in mafData files
- Added website URL to docs / comments

+++ Arduino/DIY-Flow-Bench/mafData/DELPHI_AF10118.h
@@ -1,5 +1,5 @@
 // array {mV, kgh/1000}
-float mafMapData[][2] = {
+long mafMapData[][2] = {
 {	0	,	3650	},
 {	60	,	5620	},
 {	80	,	8982	},
